outlier-detection/requirements.txt
numpy
pandas
torch
torchvision
scikit-learn
matplotlib
tensorflow
opencv-python
flask
pillow
kaggle